Revert "YSHOP2-991:  Date picker, select range should be optional (#160)"
This reverts commit b25aaef4230863912b960905b0d7d28356da2bf1.
