+++ packages/vue3/src/components/DateInput/DateInput.vue
@@ -1,27 +1,21 @@
 <script setup lang="ts">
 import { onClickOutside } from '@vueuse/core';
-import { DateUtils } from '@youcan/ui-core';
 import { computed, ref } from 'vue';
+import { DateUtils } from '@youcan/ui-core';
 import DatePicker from './DatePicker.vue';
-import DateRangePicker from './DateRangePicker.vue';
-import type { DateRangeValue } from './types';
+import type { DateInputValue } from './types';
 import SecondaryButton from '~/components/Button/SecondaryButton.vue';
 
-const props = withDefaults(defineProps<{
-  modelValue: DateRangeValue
+const props = defineProps<{
+  modelValue: DateInputValue
   disabled?: boolean
-  isSingle?: boolean
-  placeholder?: string
-}>(), {
-  isSingle: false,
-  placeholder: 'Select date range',
-});
+}>();
 
 const emit = defineEmits(['update:modelValue']);
 
 const model = computed({
   get: () => props.modelValue,
-  set: (value: DateRangeValue) => emit('update:modelValue', value),
+  set: (value: DateInputValue | null) => emit('update:modelValue', value),
 });
 
 const isDatePickerVisible = ref(false);
@@ -32,19 +26,6 @@ const toggleDatePicker = (state = !isDatePickerVisible.value) => {
   }
 };
 
-const getLabel = computed(() => {
-  if (model.value) {
-    if (model.value instanceof Date) {
-      return DateUtils.getCalendarDay(model.value, '');
-    }
-    else {
-      return `${DateUtils.getCalendarDay(model.value.start, 'Start')} - ${DateUtils.getCalendarDay(model.value.end, 'End')}`;
-    }
-  }
-
-  return props.placeholder;
-});
-
 onClickOutside(datePicker, () => toggleDatePicker(false));
 </script>
 
@@ -54,14 +35,16 @@ onClickOutside(datePicker, () => toggleDatePicker(false));
       size="sm" icon-position="right" class="input-trigger" :disabled="disabled" type="button"
       @click="toggleDatePicker()"
     >
-      <span>{{ getLabel }}</span>
+      <span v-show="model.start || model.start">
+        {{ DateUtils.getCalendarDay(model.start, 'Start') }} - {{ DateUtils.getCalendarDay(model.end, 'End') }}
+      </span>
+      <span v-show="!model.start && !model.start">Select date range</span>
       <template #icon>
         <i class="i-youcan-calendar-blank" />
       </template>
     </SecondaryButton>
     <div class="date-picker-container">
-      <DatePicker v-if="isSingle" v-show="isDatePickerVisible" ref="datePicker" v-model="model" />
-      <DateRangePicker v-else v-show="isDatePickerVisible" ref="datePicker" v-model="model" />
+      <DatePicker v-show="isDatePickerVisible" ref="datePicker" v-model="model" />
     </div>
   </div>
 </template>
