+++ packages/vue3/src/components/DateInput/types.ts
@@ -2,7 +2,7 @@ export type DayEdge = 'start' | 'end' | 'middle' | 'both' | 'none';
 
 export type DayDisplayFormat = 'numeric' | '2-letters';
 
-export interface DateRangeValue {
+export interface DateInputValue {
   start?: Date
   end?: Date
 }
